+++ x-objects/test/x/objects/post_actions_test.rb
@@ -3,6 +3,7 @@ require_relative "../../test_helper"
 module X
   class PostActionsTest < Minitest::Test
     cover Post
+    cover Objects::PostWrites
 
     def setup
       @client = FakeClient.new